refactor: move manage tokens to $lib
