+++ src/frontend/src/lib/components/manage/ManageTokens.svelte
