+++ src/frontend/src/lib/components/manage/ManageTokensMenuButton.svelte
