+++ src/frontend/src/lib/components/tokens/TokensMenu.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { Popover } from '@dfinity/gix-components';
 	import { erc20UserTokensNotInitialized } from '$eth/derived/erc20.derived';
-	import ManageTokensMenuButton from '$icp-eth/components/tokens/ManageTokensMenuButton.svelte';
 	import IconMore from '$lib/components/icons/IconMore.svelte';
+	import ManageTokensMenuButton from '$lib/components/manage/ManageTokensMenuButton.svelte';
 	import TokensZeroBalance from '$lib/components/tokens/TokensZeroBalance.svelte';
 	import Hr from '$lib/components/ui/Hr.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
